api/pyproject.toml
[project]
name = "terrabase-api"
version = "0.1.0"
description = "Terrabase API"
authors = [{name = "Tyler Walker", email = "twalker1998@gmail.com"}]
requires-python = ">=3.12"

[dependency-groups]
base = [
    "authlib==1.3.0",
    "fastapi==0.121.3",
    "fastapi-cli==0.0.16",
    "google-api-python-client==2.187.0",
    "googleapis-common-protos==1.72.0",
    "grpcio==1.76.0",
    "passlib[argon2]==1.7.4",
    "protobuf==6.33.1",
    "python-jose[cryptography]==3.3.0"
]
dev = [
    {include-group = "base"},
    "ruff==0.14.5"
]

[tool.ruff]
exclude = [
    "terrabase_api/specs"
]

[tool.pyright]
exclude = [
    "terrabase_api/specs"
]
